requirements.txt
openai>=1.0.0
firecrawl-py>=1.0.0
rich>=13.0.0
click>=8.0.0
pyyaml>=6.0
pydantic>=2.0.0
pyperclip>=1.8.0
python-dotenv>=1.0.0
streamlit>=1.30
